new change in serviceImpl and order

+++ src/main/java/com/foodservice/service/impl/DeliveryDriverServiceImpl.java
@@ -107,7 +107,7 @@ public class DeliveryDriverServiceImpl implements DeliveryDriverService {
     }
     
     @Override
-    public DeliveryDriverResponseDTO getDriverByOrder(Long orderId) {
+    public DeliveryDriverResponseDTO getDriverByOrder(Integer orderId) {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new RuntimeException("Order not found"));
